added address managent

+++ app/Http/Controllers/CartController.php
@@ -10,13 +10,13 @@ class CartController extends Controller
 {
         public function addtoCart(Request $request){
             $is_LoggedIn=(new AccessManagementController)->is_LoggedIn($request);
-            $user_Id=$is_LoggedIn;
             if($is_LoggedIn){
-              $userData=User::find($user_Id);
-              if($userData){
-                  $userData->isCustomer=1;
-                  $updateUserData=$userData->save();
-              }
+                $user_Id=$is_LoggedIn;
+                $userData=User::find($user_Id);
+                    if($userData){
+                        $userData->isCustomer=1;
+                        $updateUserData=$userData->save();
+                    }
               $Data=$request->all();
               $new_insert_array=array();
                 foreach($Data as $key=>$val)
@@ -43,23 +43,34 @@ class CartController extends Controller
         public function removeProduct(Request $request,$id){
             $is_LoggedIn=(new AccessManagementController)->is_LoggedIn($request);
             if($is_LoggedIn){
-                $Data=CartProduct::find($id);
-                if($Data){
-                    $Data->delete();
-                    return response()->json([
-                        'status'=>200,
-                        'payload'=>$Data,
-                        'message'=>'Data Deleted successfully' 
-                    ]);
-                }
-                    else{
+                $user_id=$is_LoggedIn;
+                $userData=User::find($user_id);
+                if($userData->isCustomer==1){
+                    $Data=CartProduct::find($id);
+                    if($Data){
+                        $Data->delete();
                         return response()->json([
-                            'status'=>500,
-                            'payload'=>null,
-                            'message'=>'Failed to Delete The data' 
+                            'status'=>200,
+                            'payload'=>$Data,
+                            'message'=>'Data Deleted successfully' 
                         ]);
                     }
+                        else{
+                            return response()->json([
+                                'status'=>500,
+                                'payload'=>null,
+                                'message'=>'Failed to Delete The data' 
+                            ]);
+                        }
+                }
+                else{
+                    return response()->json([
+                        'status'=>500,
+                        'payload'=>null,
+                        'message'=>'Not a customer' 
+                    ]);
 
+                }
             }
             else{
                 return response()->json([
@@ -72,24 +83,35 @@ class CartController extends Controller
         }
         public function updateQuantity(Request $request,$id){
             $is_LoggedIn=(new AccessManagementController)->is_LoggedIn($request);
-            $user_Id=$is_LoggedIn;
-            if($user_Id){
-                $Data=CartProduct::find($id);
-                $Data->quantity=$request->quantity;
-                $Data=$Data->save();
-                if($Data){
-                    return response()->json([
-                        'status'=>200,
-                        'payload'=>$Data,
-                        'message'=>'Quantity Updated  successfully' 
-                    ]);
+            if($is_LoggedIn){
+                $user_id=$is_LoggedIn;
+                $userData=User::find($user_id);
+                if($userData->isCustomer==1){
+                    $Data=CartProduct::find($id);
+                    $Data->quantity=$request->quantity;
+                    $Data=$Data->save();
+                    if($Data){
+                        return response()->json([
+                            'status'=>200,
+                            'payload'=>$Data,
+                            'message'=>'Quantity Updated  successfully' 
+                        ]);
+                    }
+                    else{
+                        return response()->json([
+                            'status'=>500,
+                            'payload'=>null,
+                            'message'=>'Failed to Update  The data' 
+                        ]);
+                    }
                 }
                 else{
                     return response()->json([
                         'status'=>500,
                         'payload'=>null,
-                        'message'=>'Failed to Update  The data' 
+                        'message'=>'Not a customer' 
                     ]);
+
                 }
             }
             else{
